backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
pytest==7.4.3
httpx==0.25.1
python-dotenv==1.0.0
requests==2.31.0  # 用于与PocketBase API通信 